use one workspace for everything browser including commands
