+++ browser/tsconfig.json
@@ -1,7 +1,6 @@
 {
 	"extends": "../tsconfig.json",
 	"compilerOptions": {
-		"declaration": false,
 		"rootDir": "./src",
 		"outDir": "./dist",
 		"paths": {
