--- commands/package.json
@@ -1,8 +1,0 @@
-{
-	"name": "command_tests",
-	"type": "module",
-	"main": "dist/mod.js",
-	"scripts": {
-		"build": "npx tsc --build"
-	}
-}
