--- commands/src/hangar.ts
@@ -1,3 +1,0 @@
-// A "hangar" or a garage for other components
-
-// This element screenshot rather than the
