--- commands/src/next_frame.ts
@@ -1,7 +1,0 @@
-export function nextFrame(): Promise<void> {
-	return new Promise(function (resolve) {
-		queueMicrotask(function () {
-			resolve();
-		});
-	});
-}
