+++ tests/src/commands/element_send_keys.ts
@@ -1,7 +1,7 @@
 import {
 	findElement,
 	elementSendKeys,
-} from "@w-lfpup/jackrabbit/commands/dist/mod.js";
+} from "@w-lfpup/jackrabbit/browser/dist/mod.js";
 
 let section = document.createElement("section");
