+++ tests/src/commands/find_element.ts
@@ -1,4 +1,4 @@
-import { findElement } from "@w-lfpup/jackrabbit/commands/dist/mod.js";
+import { findElement } from "@w-lfpup/jackrabbit/browser/dist/mod.js";
 
 let section = document.createElement("section");
