+++ tests/src/commands/find_element_from_shadow_root.ts
@@ -2,7 +2,7 @@ import {
 	findElement,
 	findElementFromShadowRoot,
 	getElementShadowRoot,
-} from "@w-lfpup/jackrabbit/commands/dist/mod.js";
+} from "@w-lfpup/jackrabbit/browser/dist/mod.js";
 
 // cannot find element in shadow root
 class MyComponent extends HTMLElement {
