+++ tests/src/commands/find_elements_from_element.ts
@@ -1,7 +1,7 @@
 import {
 	findElement,
 	findElementsFromElement,
-} from "@w-lfpup/jackrabbit/commands/dist/mod.js";
+} from "@w-lfpup/jackrabbit/browser/dist/mod.js";
 
 let section = document.createElement("section");
