+++ tests/src/commands/get_element_shadow_root.ts
@@ -1,7 +1,7 @@
 import {
 	findElement,
 	getElementShadowRoot,
-} from "@w-lfpup/jackrabbit/commands/dist/mod.js";
+} from "@w-lfpup/jackrabbit/browser/dist/mod.js";
 
 let section = document.createElement("section");
